+++ alita/utils/cmd_senders.py
@@ -1,0 +1,19 @@
+from alita import Alita
+from alita.utils.msg_types import Types
+
+
+async def send_cmd(client: Alita, msgtype):
+    GET_FORMAT = {
+        Types.TEXT.value: client.send_message,
+        Types.DOCUMENT.value: client.send_document,
+        Types.PHOTO.value: client.send_photo,
+        Types.VIDEO.value: client.send_video,
+        Types.STICKER.value: client.send_sticker,
+        Types.AUDIO.value: client.send_audio,
+        Types.VOICE.value: client.send_voice,
+        Types.VIDEO_NOTE.value: client.send_video_note,
+        Types.ANIMATION.value: client.send_animation,
+        Types.ANIMATED_STICKER.value: client.send_sticker,
+        Types.CONTACT: client.send_contact,
+    }
+    return GET_FORMAT[msgtype]
